refactor: handle unavailable features on iOS and consolidate settings UI - Refactor duplicate UI components (SettingsGroup, SettingRow, NavigationRow) into [SettingsUI.tsx] - Centralize showNotAvailableAlert in utils/alerts.ts - Replace in-app browser links with Feature not available alerts for iOS on:   - More tab items (Apps, Routing, Workflows, Insights, Support)   - Event Type Other tab (Apps, Workflows, Webhooks)   - Event Type Advanced tab (Private Links) - Remove View public page from Profile Sheet on iOS - Remove Preview actions from Event Type (List menu and Detail quick actions) on iOS - Add Learn More links to Advanced Tab settings
