+++ companion/components/event-type-detail/tabs/AdvancedTab.tsx
@@ -12,16 +12,15 @@ import {
   Modal,
   Platform,
   ScrollView,
-  Switch,
   Text,
   TextInput,
   TouchableOpacity,
   View,
 } from "react-native";
 
-import { showInfoAlert } from "@/utils/alerts";
+import { showInfoAlert, showNotAvailableAlert } from "@/utils/alerts";
 import { openInAppBrowser } from "@/utils/browser";
-import { IOSPickerTrigger } from "./IOSPickerTrigger";
+import { NavigationRow, SettingRow, SettingsGroup } from "../SettingsUI";
 
 // Interface language options matching API V2 enum
 const interfaceLanguageOptions = [
@@ -65,167 +64,7 @@ const interfaceLanguageOptions = [
   { label: "中文（台灣）", value: "zh-TW" },
 ];
 
-// Section header
-function SectionHeader({ title }: { title: string }) {
-  return (
-    <Text
-      className="mb-2 ml-4 text-[13px] uppercase tracking-wide text-[#6D6D72]"
-      style={{ letterSpacing: 0.5 }}
-    >
-      {title}
-    </Text>
-  );
-}
-
-// Settings group container
-function SettingsGroup({
-  children,
-  header,
-  footer,
-}: {
-  children: React.ReactNode;
-  header?: string;
-  footer?: string;
-}) {
-  return (
-    <View>
-      {header ? <SectionHeader title={header} /> : null}
-      <View className="overflow-hidden rounded-[14px] bg-white">{children}</View>
-      {footer ? <Text className="ml-4 mt-2 text-[13px] text-[#6D6D72]">{footer}</Text> : null}
-    </View>
-  );
-}
-
-// Toggle row with indented separator
-function SettingRow({
-  title,
-  description,
-  value,
-  onValueChange,
-  learnMoreUrl,
-  isFirst = false,
-  isLast = false,
-}: {
-  title: string;
-  description?: string;
-  value: boolean;
-  onValueChange: (value: boolean) => void;
-  learnMoreUrl?: string;
-  isFirst?: boolean;
-  isLast?: boolean;
-}) {
-  const height = isFirst || isLast ? 52 : 44;
-  const showDescription = () => {
-    if (!description) return;
-
-    const buttons: {
-      text: string;
-      onPress?: () => void;
-      style?: "cancel" | "default" | "destructive";
-    }[] = [{ text: "OK", style: "cancel" }];
-
-    if (learnMoreUrl) {
-      buttons.unshift({
-        text: "Learn more",
-        onPress: () => openInAppBrowser(learnMoreUrl, "Learn more"),
-      });
-    }
-
-    Alert.alert(title, description, buttons);
-  };
-
-  return (
-    <View className="bg-white pl-4">
-      <View
-        className={`flex-row items-center pr-4 ${!isLast ? "border-b border-[#E5E5E5]" : ""}`}
-        style={{ height, flexDirection: "row", alignItems: "center" }}
-      >
-        <TouchableOpacity
-          className="flex-1 flex-row items-center"
-          style={{ height }}
-          onPress={description ? showDescription : undefined}
-          activeOpacity={description ? 0.7 : 1}
-          disabled={!description}
-        >
-          <Text className="text-[17px] text-black" style={{ fontWeight: "400" }}>
-            {title}
-          </Text>
-          {description ? (
-            <Ionicons name="chevron-down" size={12} color="#C7C7CC" style={{ marginLeft: 6 }} />
-          ) : null}
-        </TouchableOpacity>
-        <View style={{ alignSelf: "center", justifyContent: "center" }}>
-          <Switch
-            value={value}
-            onValueChange={onValueChange}
-            trackColor={{ false: "#E9E9EA", true: "#000000" }}
-            thumbColor={Platform.OS !== "ios" ? "#FFFFFF" : undefined}
-          />
-        </View>
-      </View>
-    </View>
-  );
-}
-
-// Navigation row (with chevron)
-function NavigationRow({
-  title,
-  value,
-  onPress,
-  isFirst = false,
-  isLast = false,
-  options,
-  onSelect,
-}: {
-  title: string;
-  value?: string;
-  onPress: () => void;
-  isFirst?: boolean;
-  isLast?: boolean;
-  options?: { label: string; value: string }[];
-  onSelect?: (value: string) => void;
-}) {
-  const height = isFirst || isLast ? 52 : 44;
-  return (
-    <View className="bg-white pl-4" style={{ height }}>
-      <View
-        className={`flex-1 flex-row items-center justify-between pr-4 ${
-          !isLast ? "border-b border-[#E5E5E5]" : ""
-        }`}
-        style={{ height }}
-      >
-        <Text className="text-[17px] text-black" style={{ fontWeight: "400" }}>
-          {title}
-        </Text>
-        <View className="flex-row items-center">
-          {Platform.OS === "ios" && options && onSelect ? (
-            <>
-              {value ? (
-                <Text className="mr-2 text-[17px] text-[#8E8E93]" numberOfLines={1}>
-                  {value}
-                </Text>
-              ) : null}
-              <IOSPickerTrigger options={options} selectedValue={value || ""} onSelect={onSelect} />
-            </>
-          ) : (
-            <TouchableOpacity
-              className="flex-row items-center"
-              onPress={onPress}
-              activeOpacity={0.5}
-            >
-              {value ? (
-                <Text className="mr-1 text-[17px] text-[#8E8E93]" numberOfLines={1}>
-                  {value}
-                </Text>
-              ) : null}
-              <Ionicons name="chevron-forward" size={20} color="#C7C7CC" />
-            </TouchableOpacity>
-          )}
-        </View>
-      </View>
-    </View>
-  );
-}
+// Local components removed in favor of SettingsUI
 
 interface AdvancedTabProps {
   requiresConfirmation: boolean;
@@ -289,6 +128,8 @@ export function AdvancedTab(props: AdvancedTabProps) {
     return option?.label || "Browser Default";
   };
 
+  // handleNotAvailable replaced by global utility
+
   return (
     <View className="gap-6">
       {/* Booking Confirmation Group */}
@@ -297,6 +138,7 @@ export function AdvancedTab(props: AdvancedTabProps) {
           isFirst
           title="Requires confirmation"
           description="The booking needs to be manually confirmed before it is pushed to your calendar and a confirmation is sent."
+          learnMoreUrl="https://cal.com/help/event-types/how-to-requires"
           value={props.requiresConfirmation}
           onValueChange={(value) => {
             if (value && props.seatsEnabled) {
@@ -323,21 +165,24 @@ export function AdvancedTab(props: AdvancedTabProps) {
         <SettingRow
           isFirst
           title="Disable Cancelling"
-          description="Guests and Organizer can no longer cancel the event with calendar invite or email."
+          description="Guests and organizer can no longer cancel the event with calendar invite or email."
           value={props.disableCancelling}
           onValueChange={props.setDisableCancelling}
+          learnMoreUrl="https://cal.com/help/event-types/disable-canceling-rescheduling#disable-cancelling"
         />
         <SettingRow
           title="Disable Rescheduling"
           description="Guests and Organizer can no longer reschedule the event with calendar invite or email."
           value={props.disableRescheduling}
           onValueChange={props.setDisableRescheduling}
+          learnMoreUrl="https://cal.com/help/event-types/disable-canceling-rescheduling#disable-rescheduling"
         />
         <SettingRow
           title="Reschedule past events"
           description="Enabling this option allows for past events to be rescheduled."
           value={props.allowReschedulingPastEvents}
           onValueChange={props.setAllowReschedulingPastEvents}
+          learnMoreUrl="https://cal.com/help/event-types/allow-rescheduling"
         />
         <SettingRow
           title="Book via reschedule link"
@@ -356,6 +201,7 @@ export function AdvancedTab(props: AdvancedTabProps) {
           description="For privacy reasons, additional inputs and notes will be hidden in the calendar entry. They will still be sent to your email."
           value={props.hideCalendarNotes}
           onValueChange={props.setHideCalendarNotes}
+          learnMoreUrl="https://cal.com/help/event-types/hide-notes"
         />
         <SettingRow
           title="Hide calendar event details"
@@ -368,6 +214,7 @@ export function AdvancedTab(props: AdvancedTabProps) {
           description="Hide organizer's email address from the booking screen, email notifications, and calendar events."
           value={props.hideOrganizerEmail}
           onValueChange={props.setHideOrganizerEmail}
+          learnMoreUrl="https://cal.com/help/event-types/hideorganizersemail#hide-organizers-email"
           isLast
         />
       </SettingsGroup>
@@ -386,6 +233,7 @@ export function AdvancedTab(props: AdvancedTabProps) {
           description="Arrange time slots to optimize availability."
           value={props.showOptimizedSlots}
           onValueChange={props.setShowOptimizedSlots}
+          learnMoreUrl="https://cal.com/help/event-types/optimized-slots#optimized-slots"
         />
         <SettingRow
           title="Lock timezone"
@@ -603,10 +451,7 @@ export function AdvancedTab(props: AdvancedTabProps) {
           title="Private Links"
           onPress={() => {
             if (props.eventTypeId && props.eventTypeId !== "new") {
-              openInAppBrowser(
-                `https://app.cal.com/event-types/${props.eventTypeId}?tabName=advanced`,
-                "Private Links"
-              );
+              showNotAvailableAlert();
             } else {
               showInfoAlert("Info", "Save the event type first to configure this setting.");
             }
@@ -616,10 +461,7 @@ export function AdvancedTab(props: AdvancedTabProps) {
           title="Custom Reply-To Email"
           onPress={() => {
             if (props.eventTypeId && props.eventTypeId !== "new") {
-              openInAppBrowser(
-                `https://app.cal.com/event-types/${props.eventTypeId}?tabName=advanced`,
-                "Custom Reply-To"
-              );
+              showNotAvailableAlert();
             } else {
               showInfoAlert("Info", "Save the event type first to configure this setting.");
             }
